refactor(rollback): se refactorizado codigo relacionado a rollback
se traslada el codigo de rollback de manifest_generator.py a rollback_manager.py
